Showcase for parametric identification
The identification of the parametric audio model is working.
The requirements are that the system is left unmodified, i.e.
no compensation for equalisation or lowpass filtering of the output.

It is also important to use the simple modulation, otherwise the
identified nonlinearity is not the demodulation, but the demodulation
shifted.

Refining the aproximation also seems to work quite good.
The gaussian kernel perfors well and gives a nice smooth output.