Start of levitation
This is the first real attempt at implementing methods to
maximise the Gor'Kov Laplacian. At this stage the code
is functional and calculates results that seem reasonable.

The acoustical pressure field from the default settings give
a low pressure at the levitation point, surrounded by a high pressure
region. This is especially apparent in the xy-plane intersecting the point.