Analytical derivatives
Rework of the optimization to use the analytical derivatives of spherical
spreading. No directivity is implemented at this stage.
Includes numerous bugfixes.

The minimization converges to the vortex trap for equal weights,
but not for all staring conditions.
For large weights the algorithm might converge, but the number of
iteration is large so the code is slow.