Debug and speedup
Chage some of the debugging prints to a more compact format.
Huge speedup in the calculation of the derivatives by vectorizing.

The algorithm is still not perfect. Convergence is achieved for vortex
traps and twin traps, vortex converging fastest. Bottle traps have
poor convergence. It seems to be a local minima, but the algorithm does
not converge there unless started there. It is very sensitive to variations
in starting position.

Trials with constrained L-BFGS-B optimisation has been diffuicult to
debug since intermediate values are not easily accessible.
It is not possible to get the values back to python, all
variants have only returned the last iteration. It is however possible to
write to files using loggers, so it would be possible to reimport all
phase values from a file to see what is happening.