Extends pressure_point with gradient and sphereical grid
The 'pressure_point' is extended with two new optins:
1) It is now possible to also include the pressure gradient in the
objective. The gradient is always normalized by kr, r being the distance
to the middle of the array, to giva a more balanced contribution between
pressure and gradient. This is enabled by default, and can only be
disabled by setting all the gradient weights to 0.

2) It is possible to automatically add additional point on a spherical
grid around the main point. This has not been used with the addition
of the gradient, so the results are not known. It should however give a
larger volume where the pressure is minimized.
How to place the points on the sphere for maximim effect per point is
not known at this stage.