Changes implementation of laplacian parts
Removes the third level functions 'complex_dot' and 'phase_derivative'
in favour of s streamlined implementation. The new implementation
uses multiplications of conjugates instead of manually specifying
parts. This reduces the number of function calls and relies more on
the inherent speed of numpy.

Some old functions in the gorkov laplacian class is disbled with
'assert False' but kept for the moment.