Allow amplitudes in minimization
Change all objective functions and their jacobians to accept amplitudes
as an input. This must be in the same array as the phases for the
minimizer to understand it, and the array must be 1D.
Therefore the amplitudes can be specified after the phases as a
concatenated array. It should in principle work if the amplitudes are
specified as a second row, but this is not guaranteed for all optimizers.