Allow variable amplitudes properly
Since amplitudes must be limited to reasonable values the L-BFGS-B
algorithm is used. It seems like the algorithm is confused by limits on
the phases, so don't use that.

Implements a custom step taking algorithm that normalizes the step size
to the range of the corresponsing values. Also wraps the phase values
to [-pi, pi] and limits the amplitudes to [1e-3, 1].

The L-BFGS-B algorithm now actually works without amplitudes as well,
and has the nice benefit of being faster. (I don't know if it converges
using fewer iteration, if the algorithm is inherently faster, or if the
implementation is better optimized).

Adds starting amplitude as ones and amplitude scatter plots in the notebook.
IMPORTANT: Always start with all transducers at full power, otherwise
it might leave half of them off in a local minimum.