Reimplements cost functions as functions instead of classes
By using template functions and closures the different cost functions
are reimplemented without classes. This givea a more complact code.
It should also be easier to wrap the inner functions with performance
decorators, e.g. numba.

The new implementations calculate the function and the jacobian at the
same time. This is especially beneficial for the Gor'Kov laplacian,
since the evaluation of the function is almost free while evaluating
the jacobian.

The previous class oriented implementations of cost functions are left
untouched, but deprecated. The optimizer is modified and will not work
correctly with the old classes.