Implements calculation of forces from Gor'kov potential
Since the force is a vector, this cannot be used directly as a
optimization target. The jacobian of the force is not implemented
at this stage. If the force is to be used as a part of the objective
function, the target needs to be the force in some direction
(a weighted sum) or the magnitude. If this turns out to be of use
the jacobian can be implemented at a later stage.