Fixes sign errors in laplacian
The minus for the gradient part in the Gor'kov potential
should of course give a minus for the gradient part
of the laplacian, and thus the jacobian of the laplacian.
In practice the gradient part seems to give a very small
contribution to the total objective function. Nevertheless,
it should be implemented correctly.