Fix pressure to velocity conversion
The sign error had cancelled a dfffrent error. The equation
relating pressure gradient and velocity has a factor j, so the
squared conversion factor gives a minus sign.

All in all, these two changes does not influence the actual
results, but moving the sign to the correct place helps
readability.