Removes weights from optimizer
Since the most used cost functions have weighs, there is no need for
additional weights in the optimizer. If a pre-made cost funcion must be
weighted after creation, it is very easy to do with a lambda. If this
becomes a frequenct task, a 'weghted_cost_function' function generator
could be added to created a scaled cost function.