Major rehaul of cost function weight argument
Changes the usage of the cost function weight argument to some extent.
If not passed (or `None`), the created function will return individual
parts of the cost functions, e.g. Uxx, Uyy, Uzz. If `weight` is passed
as an iterable, the parts will be weigted by said weights. The output
will instead be the summed parts and the resulting jacobian.

This changes the effect of the weights for the laplacian, so that the
"normal" weights should be negative for the laplacian parts.

The name of `gorkov_force` is changed to `gorkov_derivative` to better
represent the mathematical definitions. This is also extended with a
jacobian. The usefulness of the jacobian can be argued, since direct
min/maximization of the forces is probably not of any practical use.