Implements second order stiffness
Building on the expressions for the second order force, an expression
for the second order stiffness is imlemented. The implementation works
in a similar way as the previous Gor'kov Laplacian, that is, it is
possible to specify four weights in order to also minimize the pressure
at the same time. The difference is that the stiffness should be
minimized (maximum negative value) while the Gor'kov Laplacian should
be maximized (maximum positive value) for the forces to converge.