Removes squared pressure from cost functions
It will only be confusing to return squared pressure from some functions
but not from others, so the small performance boost (~7%) from
caculating it at the same time is not worth it.
Instead of weighting with four weights to create a trap, use an
additional pressure_null cost funciton with just a single weight. If we
want to regain some performance we could make it possible to only use
the 0th order "derivative" in the presssure_null.