Cost function generator
Since a lot of the const function code is the same in many functions, it
makes sense to have a common implementation. The only difference between
many cost functions is how to caculate values and jacobians from the
derivatives, which is not the only part of a cost function which needs
to be created in order to have a new implementation.