Separates pressure_null
The pressure_null costfunction was specifically designed to ceate quiet
zones. Currently the cost_functions module is moving away from pre-made
featues in favour of lower level fields.
The pressure_null function is replaces by a `pressure` function, and a
`velocity` function, which works as the components of the previous function.
This is much easier to understand since no "magic" is going on depending
on how many weights are passed in.