Receiver position shape check
If receiver positions is passed in with shape (...,1) everyting was
still continuing, which should not be the case. Additional checks to make
sure that the last dimention is of length 3 are now in place. It is still
not fool-proof, since a 3x3x3 mesh of points will always pass all tests
regardless of how the data is organized.