More examples
Moves the description of the examples to comments within the python files.
It looks a little less nice in the actual docs, but the files are then
standalone examples files.

Makes the docs depend on the example files in the make commands, so that
the documentation will rebuild if the examples are changed.