New DragonflyArray class
It is convenient to have access to the signatures when working with a
dragonfly array. this class is a simple subclass of rectangular array,
which bypasses the normal generation routine by loading the hard coded
positions from the hardware module. The rotation code is copied from
the rectangular array class.