Bugfix in timer
There was a small bug which occurred when rate was set but no states
were loaded. Since the size of the states variable was then zero, there
was a modulus with zero problem in the timer thread.

Annoyingly difficult to pinpoint since the application just exited
without any error message at all...