Bugfix in TCPArray
The c++ side expects one message per state, not one huge message
with all states. it is a lo t easier to change the python side to send
multiple messages. It should also be more robust since the risk
of flooding the TCP connection should (hopefully) be lower with
multiple small messages?

The `close` function now properly checks if the c++ process is still
running and sends the exit signal if it is.
The python side socket and connection can safely be closed multiple
times, so no check is needed.