Transpose of the signature generators
Rectangular array has some signature calculators which are now transposed.
The new implementations are also much nicer. In particular the twin
signature is better. The old implementation had problems with the two
halves switching places at certain angles, which gives a sharp polarity
switch in paths. This is not the case in the new implementation.

NOTE: The twin signature function has a changed output for certain inputs.