Fixes file read bug in c++
Since variable lengths arrays are not allowed on windows compilers,
this was changed but not properly tested some time ago.
As a result the states in the read file was read one value at the time,
and stored only in the first transducer for each state with 256 times
too many states.

This fixes the file reading so that the previously written files can
actually be read properly.

Note: only tested on osx, so windows might break again...