Visualization improvements
Sets a constant reference value for sound velocity level. This is more
in line with the normal engineering standards.

Improves the transducer visualization for easier plotting of signatures,
as well as easier extension if needed.

Creates a multiple field figure generator, which can be used to easier
plot multiple fields in a single plot and rapidly switch between them.