Renames algorithm requirements and arguments
Changes the  the algorithm arguments to the same names as the specified
requirements. This allows a more robust compilation of call arguments
from a larger list by utilizing the `requires` dict and selecting the
appropriate keys in the larger dictionary of call arguments.

Renames the requirement to something more descriptive. `pressure_orders`
is not very nice because all possible requirements could have an "order",
but "derivs" is more descriptive.