Implements new algorithm wrapper classes
The new classes are designed around emulating numerical types, i.e.
they support a set of arithmetical operation.
This greatly simplifies the creation of the desired cost functions for
optimization, as well as aligns the top-level api with the theoretical
formulations.

There is some additional overhead due to method lookups, but the package
is not really designed for uttermost speed in execution anyhow.

Note that almost the entire package is broken at this point, since
everything is build on top of the algorithms...