Splits TransducerModels
Some functionality is not really part of a general transducer model, but
only part of the point source models. A new class `PointSource` for
said models is created, collecting the relevant implementations.

The advantage at this point is that the `PlaneWaveTransducer` does not
have a lot of methods which does not make any sense at all.
In the long run we have the possibility to implement another layer of
fallbacks in the top level transducer model, e.g. using numerical
calculations with only a greens function to calculate some spatial
information.