Bugfix for algorithms with non-keyword arguments
When taking input array as a positional argument but passing it to the
underlying func as a keyword argument the array argument will be placed
as a keyword argument. In combination with `*args` the end result is
that the first element in args will also be passed as array, which is of
course not possible.

This change is necessary to enable positional arguments in algorithms,
at the cost of fixing the array as the first positional argument to all
algorithms.