Streamlines optimization sequences
Since we are lo longer passing a list of cost functions for a single
minimization run, we can instead interpret an iterable as a minimization
sequence directly. This also makes the calling of the cost function
easier, since we do not have to deal with multiple functions inside the
minimizer.