Fixes the usage of PersistentFieldEvaluator in loops
It is sometimes more efficient to use the persistent evaluator win e.g.
kinetic simulations, since the same point might be used for more than
one evaluation. If the calculation point was modified in place the
evaluator's last position was also modified. Solved by copying the input
position in the evaluator.