Enables algorithms with arbitrary dimensions
Some algorithms calculate values which are inherently with a different
number of dimensions than 1, e.g. the Gor'kov potential which have 0.
Some ideas in the future might need more dimensions, e.g. matrices.
This adds a mandatory specification for algorithms for how many
dimensions the output will have (on top of the number of dimensions for
the coordinates).

As a result, some of the weighting stuff needs tom be more flexible in
the implementation to handle weights with different number of dimensions.
The caveat is that the weight needs to be specified with a compatible
number of dimensions to work. If there are too few dimensions, extra dims
are added similar to normal broadcasting rules, i.e. prepended.
Too many dimensions is not possible to fix automatically since the
meaning if not very well defined.

The tests are modified slightly to work with the new output format of
the two 0 dim algorithms. The major change in the test suite is that
the pressure_squared_magnitude algorithm is no longer part in testing
the algorithm structures, since it cannot handle the 1-dim weights.