Implements the last third order pressure spatial derivative
Until not, the `xyz` derivative of the pressure was not needed.
This commit both implements the derivative in the transducer models and
makes sure that the storage descriptions includes it.

The tests are updated to include the new derivative in the transducer
model tests and the array tests.
There is also a internal consistency check to make sure that the
implemented transducer models give the same results using a manual
finite difference scheme, and the implemented functions.