Merges the signature functions
Instead of implementing a separate signature function for each signature
type they are not merged in one function and controlled with an argument.
Keeping separate functions turned out to be very difficult when dynamic
class generation is not allowed. Using arguments allows us to accept any
signature type in any array, and then implement ways to find the
requested signature even if it is not in the namespace of the current
class.