Re-implements DoublesidedArray
The dynamic class creation, while cool and neat, gives an implementation
which cannot be pickled. This is annoying since we would need custom
ways to save the arrays.

This new implementation simply creates the required transducer positions
by mirroring the ones in a singlsided array. Since the signatures,
the only thing we actually used the singlesided class for, are now
possible to access through the unified signature function, we don't
need to subclass the singlesided array.