Simplify array string formating
Especially the doublesided array string formating was a bit too messy.
Not the array subclasses simply add extra print arguments in a dict,
and they will be used for a replace if possible.

Doublesided array will simply print the singlesided array used to create
the grid as the formated output.