Updates doublesided array test
The new implementation of doublesided array has some differences to the
old one. For "normal" usage where the array is nicely parallel with the
axes, the new implementation has simply changed the order of the elements.
For cases when the array normal is not zero in any direction the old
implementation returned a doublesided array where the top half was not
a simple mirror of the bottom half. The validity of the new implementation
was confirmed bu checking that each element in the top half is separated
with the correct distance in the direction of the normal. This was not
always true for the old implementation.