New AlgorithmImplementation class
This will solve the problems with pickleability in the long run. We need
a custom `__getnewargs__` method to tell pickle that the array is needed
when creating the objects, not only when initializing them. This is
since the __new__ method does not return an object of the called class,
so there is no automagic calling of __init__. Therefore __init__ is
manually called in __new__, both for the implementations and the wrappers.