Re-implements materials as classes
It turns out that the previous implementation with a single `Material`
class and instances for each material did not pickle nicely. The problem
was the `update_properties` function of `Air`, which was still pseudo-bound
to the instance before pickling. I suppose that the method would have
broken if used in a new python kernel.

This new implementation uses a metaclass for the materials, and classes
for each material. This allows for normal classmethods to update
class properties, and the `Material` metaclass acts like the class
of the classes (it is). A pickled material class will simply point to
the class.

Caveat: it is not possible to store a change in material properties!