Adds requirements attributes to algorithm classes
An issue was discovered where instances of the same algorithm could not
have different requirements, since the are stored in the class-level
method, not the instance level method. This is a problem for spherical
harmonics based calculations, as well as the vector algorithms.

We will need to rework the while requirements protocol to store them as
class attributes or instance attributes as needed, and not as a property
of the `calc_...` methods.