Separate out algorithm methods
There are several methods in `Algorithm` which do not make sense for
`AlgorithmPoint` etc. In order to remove the surplus methods of the
higher level classes, we need to move the common ones to a lower level
class. To track if a certain class is bound/unbound, algorithm/cost we
add two attributes and a property to stack them. This is simpler than
checking for inheritance since we don't actually want to inherit the
methods.