Moves caching of spatial structures to AlgorithmBase
This is needed here and there, so it's better if it's common to all.
We should not have any problems with reusing the cache when the position
is changes, since we will have a new instance of the algorithm class.
Changed requirements might require a flush of the cache, but that's
best handled while updating the requirements.