Move Algorithm wrapping of implementations to metaclass
The previous wrapping of AlgorithmImplementation inside an Algorithm
required some workarounds for pickle. Moving the custom object creation
to a metaclass solves this, since pickle is not actually calling the class.

It would be possible to implement a 'nowrap' parameter to prevent the wrapping,
but at this point it is not needed.