Fix tests and examples
Replaces PressureMagnitudeSquared and VelocityMagnitudeSquared if all
tests and examples.

In the tests of algorithm calculations the VelocityMagnitudeSquared
test case is removed. Since it is now a MagnitudeSquaredAlgorithm it
will not behave the same way as the others when used to subtract a target.
This is not unexpected, and was not a reasonable thing to have as a cost
function to start with.