Moves documentation to module level docstrings
It is nice if the source for the documentation is actually included
inside the source files, instead of the sphinx doc files.
This way the overall documentation is also available from the interpreter,
except for some more general stuff.

In the long run we might move the current "introduction" and "package"
rst files to live inside the main __init__ file. See for example how the
numpy docs are organized: almost all the documentation is inside the
actual source files, and there is probably just some "small" sphinx
stuff somewhere to link it all.