Bugfix for medium properties in transducers
Due to local storage of the wavenumber k in transducer models, changing
the medium properties after creating a transducer did not update the
wavenumber until the transducer frequency was reset. Effectively,
existing transducer would not update to respect the new medium properties.

This removes the stored k value, in favor of always calculating it using
omega and the wave speed.

Due to how pytest works we need to add lines to change the air properties
to fixed values inside each test file. What happened was that the test
files with "early names", e.g. algorithm_calculations is imported first,
creating a transducer model and calculating some derivatives.
Later on other test modules are imported, changing the medium properties.
When the tests in `alorithm_calculation` are executed with the changed
properties, the results will not match. Before this change there was no
issue since the bug prevented the transducer to use the new medium
properties.