Simplifies analytic expressions for radiation force coefficients
The coefficients for the radiation force expression can be simplified in
a similar way to the radiation force curl. This reduces the number of
statements both in the init but also in the evaluations.
It is also more interesting from an scientific point of view since it
highlights that the pressure coefficient is mostly dependent in the
compressibility and the velocity coefficient is only dependent on the
densities.