Merges RadiationForceCurl methods
Using the new simplifications for RadiationForce, it is apparent how
to use the same implementation for the curl as well. It is very likely
that if the coefficients are changed due to e.g. new scattering models,
the curl should change in the same way. Therefore it is only reasonable
to have a single implementation and use that everywhere.