Rename materials with lowercase names
In order to make the material properties pickleable they cannot be
classes, so we will need instances. It will be convenient to have
both classes and instances for each material, so we will use lowercase
names for the instances of the camel case classes.
To prepare the rest of the codebase for the change, we rename all
current material implementations using lowercase names.