Streamlines handling of requirements in FieldBase
There is no need for the other field wrappers to request the spatial
structures on their own. Field base is doing all the work anyhow, so it
makes sense that everything has a single entry point.

Implementing this in `__call__`: This would have made the most sense,
but there is no nice way to handle the correct class structure for this,
other than `FieldBase.__call__(self,...)` which is not very nice.

Implementing this in `FieldRequrement`: This will never be able to handle
caching, since most of the requirements are class variables. This means
that the subclasses of `FieldBase` would still have to call some method
on `FieldBase`, so it doe not really make any difference.