Moves FieldRequirement to internal class in FieldBase
FieldBase is the class where all the logic for evaluating requirements
is handled. It makes sense that a class for storing requirements lives
inside FieldBase.

FieldBase now has three methods to handle this logic.
`_evaluate_requirements` takes the complex transducer amplitudes and
an optional position. This is the method which should be called from
subclasses when they need the requirements.
`__evaluate_requests` asks the transducer array to calculate the data
we need to evaluate the requirements, as well as some additional steps
needed for special requirements.
`__requests_needed` is a small helper function which only finds which
requests will be needed from the array.