Simplifies the spherical harmonics forces
Since we calculate the x and y forces on one go, the performance gain
form indexing the arrays first and calculating later is so small that it
is not worth it. There is only a single array which is only indexed
twice. This implementation is more straight forward, and should allow
for less code to be needed to implement the jacobians the same way.