Updates travis build for python 3.8
Travis-CI claims to be stable on python 3.8, so we should include tests
for that now. Sometime during the last year travis moved all builds over
to xenial and changes something so that sudo is no longer required.
This simplifies out build matrix since everything runs in the same
environment, only the python version changes.