Simplifies `greens_function` and renames to `pressure`
This function is very rarely used, but sometimes it is useful.
The previous implementation forced subclasses to implement both pressure
derivatives and greens_function. This is not ideal and leads to code
duplication. Now, the `pressure` function (replacing `greens_function`)
just calls `pressure_derivs` and gets only order zero derivatives.