Moves automatic mesh management to `Trace`
Implements a `meshproperty` decorator as a simple way to ensure that
the mesh is updated on property changes.
Allows for a "preprocesing" stage which is evaluated on property set,
and a "postprocessing" which is evaluated on property get.

Preprocessing is typically used to e.g. normalize the values.
Ii have not yet found any use for the postprocessing, but at some point
it might be useful.