Enables gravity compensation in visualizations
Since the visualizations are somewhat integrated now, it will be convenient
to have a shorthand to include gravity in the calculations. Instead of
trying to get hold of the sphere properties in the visualizers, all
field classes where gravity will have a defined value will calculate and
store `mg`. It is somewhat hidden from regular users since it will be
accessed as `field.field.mg`, and most people would not look for it there.