Moves handling of array positions and normals to properties
There's a number of steps we should always do when setting the transducer
array positions and normals. Perhaps the most important is to actually
normalize the normals. This also checks that the positions and normals
have the correct shape, and sets the number of transducers automatically.
The number of transducers are now a non-settable property, which makes
more sense.