Implements Array visualization
Gets the plotly traces of multiple traces and returns all of them in a
figure. Creates updatemenus to toggle between multiple fields and
transducer visualization combos. The toggling is unfortunately somewhat
slow to render in the browser, but still faster than replotting the
entire thing.

I tried to redraw the data and colorscale of the transducer traces
instead of just drawing multiple ones and toggling, but it did not
render faster so there was no apparent gain.