Removes the notion that a transducer array has a state
The properties `phases` `amplitudes`, and `complex_amplitudes` are
removed from `TransducerArray`. The entire package has moved away from
using the transducer array as a holder for the state. Now it is only a
container for the transducer objects.

This way there is no ambiguity
of what properties will actually be used in calculations, or how the
relation between the amplitude and phase form corresponds to the
complex form.

A few convenience functions are added in the utils module, mostly as
some sort of documentation of the conventions used for the rest of the
code.