Implements deferred mesh generation for visualization traces
Since we auto-generate the visualizers on array initialization they will
always be available. This came at the cost of always generating the mesh
for the transducer visualization, even in scripts where no visualization
was ever done.
This change to the MeshTrace class makes sure that the mesh is not
created until it is actually needed, saving time and memory.