Updates to visualization
Adds nicer string formating for a lot of the text in the figures.
When the TeX rendering of text is not available, we have to resors to
manual application of html tags.

Adds an option to project the array visualizations on a given axis plane.
Userful for static "2d"  plots.