Moves more methods to NormalTransducerArray
Signatures are at least properly implemented for all variants of the normal
transducer array.
The previous implementation of RectangularArray using a classmethod to generate the grid is not
really needed.  It's easier to just generate the positions in the
initializer, since the rotations etc. are taken care of in the super class.