Moves DragonflyArray class to hardware module
It makes more sense to have both the classes for simulating existing
arrays and the classes for driving them in the same place. This uses
a "new" way to get the grid for the dragonfly array which uses the
positions in the array grid and the transducer spread, instead of
manually hardcoding each element.