Add option to exclude center for cone visualization
In many cases the vector in the center of a vector field visualization
is mostly in the way. This adds the option (default on) to remove that
central point from the mesh. This seems to increase the readability
of the force visualizations, since the force in the center is mostly
zero anyhow.