Moves calculation of jacobians
This moves the calculation of jacobians to just get them from the
`values_jacobians`. This is needed since when transformers are used it
is not possible to calculate the jacobians without also calculating the
values. It is quite rare that the jacobians are needed without the
values, so the performance difference should be minor.