Implements easier stacking of fields
A convenient function to stack multiple fields is added to the
wrappers module. This allows a user more specific control over how the
fields will be nested, without having to deal with the class instantiation
directly. It also serves as a simple way to stack fields internally in
the module.