Updates optimization module for new wrappers
Adds automatic conversion to cost function instead of field. This is
where a user would be notified if the cost function is not properly
defined.
Changes some minor checks to determine if the optimization is sequenced
or not. This will no longer work with callables that are not from the
wrappers.