Adds simple divergence and curl classes for spherical harmonics force
These quantities are often used in cost functions, so it makes sense to
implement it in the package instead of re-implementing this every time
it is needed. The implementation is not particularly efficient, since it
starts by calculating the full gradient matrix. It is possible to reduce
the computations needed for the two quantities by a lot...